First pass in restructuring the project
