+++ docs_src/impl/1. Library Architecture.md
@@ -5,13 +5,14 @@ The epiworld library is structured around a modular architecture that models pop
 
 | Directory      | Description                                                        |
 | -------------- | ------------------------------------------------------------------ |
-| `include/`     | Header files for the core of `epiworld`, and vendored libraries    |
+| `include/`     | Header files for the core of `epiworld`
+| `vendor/`      | Header files for the vendored libraries
 | `examples/`    | Example models and usage demonstrations                            |
 | `tests/`       | Unit or integration tests                                          |
 | `benchmark/`   | Performance benchmarks                                             |
 | `docs/`        | User documentation, model descriptions, perhaps architecture notes |
 | `epiworld.hpp` | The pre-built, pre-packaged "monoheader" for epiworld.             |
-There are files at the root of the repository (e.g. `Makefile`, `Doxyfile`), but these should be mostly self explanatory, and/or stock. Usage is header-only, with template-based implementations in `epiworld.hpp` (or in files under `include/`). The library depends only on the C++ standard library (since it is a header-only template library); the `cxxopt` and `catch2` libraries are used in the examples and unit tests (which are not required under normal operation).
+There are files at the root of the repository (e.g. `Makefile`, `Doxyfile`), but these should be mostly self explanatory, and/or stock. Usage is header-only, with template-based implementations in `epiworld.hpp` (or in files under `include/`). The library depends only on the C++ standard library (since it is a header-only template library); the `catch2` library is used in the unit tests (which are not required under normal operation).
 
 ## Bones and Meat
