+++ include/README.md
@@ -1,12 +1,6 @@
 # Source code
 
-Although `epiworld` is a header-only C++ library, we do make use of other
-cool projects in some of the examples and for testing. The
-[`cxxopts`](https://github.com/jarro2783/cxxopts) library provides a smooth
-interface for using options in your C++ program.
-[`catch2`](https://github.com/catchorg/Catch2) is a great C++ unit-tests
-framework that we have use in [other projects](https://github.com/USCbiostats/barry).
+Although `epiworld` is a header-only C++ library, we do make use of [`catch2`](https://github.com/catchorg/Catch2); a great C++ unit-tests framework that we have use in [other projects](https://github.com/USCbiostats/barry).
 
 Although you are free to use `epiworld/`, we also have a single-header
 version that incorporates all its components [here](../epiworld.hpp).
-
