+++ tests/main.cpp
@@ -1,3 +1,3 @@
 #define CATCH_CONFIG_MAIN
-#include "../include/catch2/catch.hpp"
+#include "../vendor/catch2/catch.hpp"
